requirements.txt
Flask
google-cloud-firestore
google-cloud-storage
google-cloud-logging
google-cloud-error-reporting
google-cloud-pubsub
gunicorn
six
pillow
